Improving code shape, refactoring notifications
